requirements.txt
matplotlib>=3.9.0
pystray>=0.19.0
Pillow>=10.0.0
tkcalendar>=1.6.1
